+++ custom_components/supernotify/media_grab.py
@@ -249,8 +249,11 @@ def select_avail_camera(hass_api: HomeAssistantAPI, cameras: dict[str, Any], cam
     if preferred_cam and CONF_CAMERA not in preferred_cam:
         preferred_cam[CONF_CAMERA] = camera_entity_id
     if preferred_cam is None:
-        # assume unconfigured camera available
-        return camera_entity_id
+        # no alternatives for an unconfigured entity, whether camera or image, so it's just
+        # a matter of whether it's known to be unavailable - assume it's fine if it has no state
+        if camera_available(hass_api, {CONF_CAMERA: camera_entity_id}, non_entity=True):
+            return camera_entity_id
+        return None
     if camera_available(hass_api, preferred_cam):
         return camera_entity_id
 
@@ -273,7 +276,7 @@ def select_avail_camera(hass_api: HomeAssistantAPI, cameras: dict[str, Any], cam
                 _LOGGER.info(
                     "SUPERNOTIFY Selecting alt camera %s with no known entity for %s", alt_cam[CONF_CAMERA], camera_entity_id
                 )
-            return alt_cam[CONF_CAMERA]
+                return alt_cam[CONF_CAMERA]
 
     return None