+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 ### Other
 - HomeAssistant compatibility moved to 2026.9.3
 - Fix handling of target specific data, that could lead to targets showing up in other envelopes
-- Internal
+- If a camera or other image source not available, the mobile push is text only rather than showing with a broken image
+
+### Internal
   - `EnityCategory` renamed to `TargetEntityCategory` to avoid naming clash with Home Assistant term
 
 ### Scenario and Recipient Switches